deps: burn down layer warnings
